requirements.txt
fastapi
uvicorn
keras
tensorflow
pillow
numpy
python-multipart
